# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能知识库平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/models/providers": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Providers Endpoint",
        "operationId": "list_providers_endpoint_api_v1_models_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Providers Endpoint Api V1 Models Providers Get"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Upsert Provider Endpoint",
        "operationId": "upsert_provider_endpoint_api_v1_models_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upsert Provider Endpoint Api V1 Models Providers Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/models/providers/{code}": {
      "patch": {
        "tags": [
          "models"
        ],
        "summary": "Patch Provider Endpoint",
        "operationId": "patch_provider_endpoint_api_v1_models_providers__code__patch",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Provider Endpoint Api V1 Models Providers  Code  Patch"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Provider Endpoint",
        "operationId": "delete_provider_endpoint_api_v1_models_providers__code__delete",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Provider Endpoint Api V1 Models Providers  Code  Delete"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models Endpoint",
        "operationId": "list_models_endpoint_api_v1_models_get",
        "parameters": [
          {
            "name": "provider_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Code"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Models Endpoint Api V1 Models Get"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create Model Endpoint",
        "operationId": "create_model_endpoint_api_v1_models_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Model Endpoint Api V1 Models Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelCreate": {
        "properties": {
          "provider_code": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "deepseek",
              "ollama",
              "custom",
              "dashscope"
            ],
            "title": "Provider Code"
          },
          "model_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Model Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "chat",
              "embedding",
              "rerank"
            ],
            "title": "Kind"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "明文密钥，仅写入"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4096.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "distance": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "cosine",
                  "l2",
                  "inner_product"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N"
          }
        },
        "type": "object",
        "required": [
          "provider_code",
          "model_name",
          "kind"
        ],
        "title": "ModelCreate"
      },
      "ModelProviderCreate": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "deepseek",
              "ollama",
              "custom",
              "dashscope"
            ],
            "title": "Code"
          },
          "display_name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Display Name"
          },
          "base_url": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Base Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "display_name",
          "base_url"
        ],
        "title": "ModelProviderCreate"
      },
      "ModelProviderUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ModelProviderUpdate"
      },
...
      "APIErrorResponse": {
        "type": "object",
        "title": "APIErrorResponse",
        "description": "所有非健康检查 API 的统一错误响应。",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API 前缀"
    }
  ]
}